requirements.txt
playwright
python-dotenv
loguru
beautifulsoup4
lxml
curl-cffi
